separate asset loading and animating into their own classes
