+++ WallyMapSpinzor2.Raylib/src/Gui/Properties/BackgroundProps.cs
@@ -48,9 +48,9 @@ public partial class PropertiesWindow
                     }
                 });
             }
-            if (data.Canvas is not null)
+            if (data.Loader is not null)
             {
-                Texture2DWrapper texture = data.Canvas.LoadTextureFromPath(Path.Combine(backgroundDir, b.AssetName));
+                Texture2DWrapper texture = data.Loader.LoadTextureFromPath(Path.Combine(backgroundDir, b.AssetName));
                 rlImGui.ImageSize(texture.Texture, new Vector2(200 * (float)(texture.Width / texture.Height), 200));
                 (b.W, b.H) = (texture.Texture.Width, texture.Texture.Height);
             }
@@ -89,9 +89,9 @@ public partial class PropertiesWindow
                     propChanged = true;
                 }
             }
-            if (data.Canvas is not null && b.AnimatedAssetName is not null)
+            if (data.Loader is not null && b.AnimatedAssetName is not null)
             {
-                Texture2DWrapper texture = data.Canvas.LoadTextureFromPath(Path.Combine(backgroundDir, b.AnimatedAssetName));
+                Texture2DWrapper texture = data.Loader.LoadTextureFromPath(Path.Combine(backgroundDir, b.AnimatedAssetName));
                 rlImGui.ImageSize(texture.Texture, new Vector2(200 * (float)(texture.Width / texture.Height), 200));
                 if (texture.W != b.W || texture.H != b.H)
                     _warningText = "AnimatedAssetName image is not the same size as the AssetName image. This can lead to the image displaying incorrectly";
